fix(cmd/paratime): Revamp paratime stats
Removed primary invoked statistics as primary workers are expected
to be continuously active.
Introduced new statistics to monitor node's block proposals as primary
or backup proposer.
Reorganized columns and sorted rows within the entity statistics table
to improve clarity.
Fixed a bug that led to inaccurate statistics when the discrepancy
event for the first finalized block occurred before the start block.
Resolved a bug that caused events from the start block to be skipped.